refactor(audit): split emitter wire contracts
